refactor(agents): move agent settings endpoints to a dedicated controller
Extract the revision history, restore, publish and archive endpoints from
AgentsController into a new AgentSettingsController, introduce AgentSettingsDto
for revision payloads, and move agent DTO mapping into agent.mapper.ts with
support for draft revisions.
