+++ apps/api/src/domains/agents/agents.controller.ts
@@ -1,6 +1,4 @@
 import {
-  type AgentDto,
-  AgentSettingsRoutes,
   type AgentSubAgentDto,
   AgentSubAgentsRoutes,
   AgentsRoutes,
@@ -13,13 +11,10 @@ import {
   Controller,
   Delete,
   Get,
-  NotFoundException,
-  Param,
   Patch,
   Post,
   Put,
   Req,
-  UnprocessableEntityException,
   UseGuards,
   UsePipes,
 } from "@nestjs/common"
@@ -33,14 +28,12 @@ import { ResourceContextGuard } from "@/common/context/resource-context.guard"
 import { CheckPolicy } from "@/common/policies/check-policy.decorator"
 import { ZodValidationPipe } from "@/common/zod-validation-pipe"
 import { TrackActivity } from "@/domains/activities/track-activity.decorator"
-import { extractAgentSettingsUpdateFields } from "@/domains/agents/settings/agent.settings.functions"
-import type { AgentSettings } from "@/domains/agents/settings/agent-settings.entity"
 // biome-ignore lint/style/useImportType: Required at runtime for NestJS DI
 import { AgentSettingsService } from "@/domains/agents/settings/agent-settings.service"
 import { JwtAuthGuard } from "@/domains/auth/jwt-auth.guard"
 import { UserGuard } from "@/domains/users/user.guard"
-import type { Agent } from "./agent.entity"
 import { AgentGuard } from "./agent.guard"
+import { toAgentDto, toAgentWithDraftDto } from "./agent.mapper"
 // biome-ignore lint/style/useImportType: Required at runtime for NestJS DI
 import { AgentsService } from "./agents.service"
 import type { AgentSubAgent } from "./sub-agents/agent-sub-agent.entity"
@@ -108,12 +101,22 @@ export class AgentsController {
     })
     const results = await Promise.all(
       agents.map(async (agent) => {
-        const agentSettings = await this.agentSettingsService.getLast({
+        const currentAgentSettings = await this.agentSettingsService.getLast({
+          connectScope,
+          agentId: agent.id,
+        })
+
+        const draftAgentSettings = await this.agentSettingsService.getLast({
           connectScope,
           agentId: agent.id,
           includesDraft: true,
         })
-        return toAgentDto({ agent, agentSettings })
+
+        const hasDraft =
+          draftAgentSettings.isDraft && currentAgentSettings.id !== draftAgentSettings.id
+        if (hasDraft)
+          return toAgentWithDraftDto({ agent, currentAgentSettings, draftAgentSettings })
+        else return toAgentWithDraftDto({ agent, currentAgentSettings })
       }),
     )
     return { data: results }
@@ -138,138 +141,6 @@ export class AgentsController {
 
     if (!agent) {
       throw new Error("Agent not updated")
-    }
-    return { data: { success: true } }
-  }
-
-  @Get(AgentSettingsRoutes.getAll.path)
-  @CheckPolicy((policy) => policy.canUpdate())
-  @AddContext("agent")
-  async getAllHistory(
-    @Req() request: EndpointRequestWithAgent,
-  ): Promise<typeof AgentSettingsRoutes.getAll.response> {
-    // Drafts are included so the history timeline can show the pending draft revision and diff
-    // it against the last published one.
-    const agentSettings = await this.agentSettingsService.getAll({
-      connectScope: getRequiredConnectScope(request),
-      agentId: request.agent.id,
-      includesDraft: true,
-    })
-    const results = agentSettings.map((as) => {
-      return toAgentDto({ agent: request.agent, agentSettings: as })
-    })
-    return { data: results }
-  }
-
-  @Post(AgentSettingsRoutes.restoreOne.path)
-  @CheckPolicy((policy) => policy.canUpdate())
-  @AddContext("agent")
-  @TrackActivity({ action: "agent.update", entityFrom: "agent" })
-  async restoreOne(
-    @Req() request: EndpointRequestWithAgent,
-    @Param("revision") revisionParam: string,
-  ): Promise<typeof AgentSettingsRoutes.restoreOne.response> {
-    const revision = Number(revisionParam)
-    if (!Number.isInteger(revision) || revision < 1) {
-      throw new UnprocessableEntityException(`Invalid revision "${revisionParam}"`)
-    }
-
-    const connectScope = getRequiredConnectScope(request)
-    const targetSettings = await this.agentSettingsService.get({
-      connectScope,
-      agentId: request.agent.id,
-      revision,
-    })
-    if (!targetSettings) {
-      throw new NotFoundException(
-        `Revision ${revision} not found for agent with id ${request.agent.id}`,
-      )
-    }
-
-    await this.agentsService.updateAgent({
-      connectScope,
-      agentId: request.agent.id,
-      fieldsToUpdate: extractAgentSettingsUpdateFields(targetSettings),
-    })
-
-    return { data: { success: true } }
-  }
-
-  @Post(AgentSettingsRoutes.publishOne.path)
-  @CheckPolicy((policy) => policy.canUpdate())
-  @AddContext("agent")
-  @TrackActivity({ action: "agentSettings.publish", entityFrom: "agentSettings" })
-  async publishOne(
-    @Req() request: EndpointRequestWithAgent,
-    @Body() { payload }: typeof AgentSettingsRoutes.publishOne.request,
-    @Param("revision") revisionParam: string,
-  ): Promise<typeof AgentSettingsRoutes.publishOne.response> {
-    const revision = Number(revisionParam)
-    if (!Number.isInteger(revision) || revision < 1) {
-      throw new UnprocessableEntityException(`Invalid revision "${revisionParam}"`)
-    }
-
-    const connectScope = getRequiredConnectScope(request)
-    const targetSettings = await this.agentSettingsService.get({
-      connectScope,
-      agentId: request.agent.id,
-      revision,
-    })
-    if (!targetSettings) {
-      throw new NotFoundException(
-        `Revision ${revision} not found for agent with id ${request.agent.id}`,
-      )
-    }
-    const { revisionName, revisionDesc } = payload
-    const updated = await this.agentSettingsService.publish({
-      connectScope,
-      agentId: request.agent.id,
-      revision,
-      revisionName,
-      revisionDesc,
-    })
-    if (!updated) {
-      throw new UnprocessableEntityException(
-        `Unable to publish revision ${revision} for agent with id ${request.agent.id}`,
-      )
-    }
-    return { data: toAgentDto({ agent: request.agent, agentSettings: updated }) }
-  }
-
-  @Post(AgentSettingsRoutes.archiveOne.path)
-  @CheckPolicy((policy) => policy.canDelete())
-  @AddContext("agent")
-  @TrackActivity({ action: "agentSettings.archive", entityFrom: "agentSettings" })
-  async archiveOne(
-    @Req() request: EndpointRequestWithAgent,
-    @Param("revision") revisionParam: string,
-  ): Promise<typeof AgentSettingsRoutes.archiveOne.response> {
-    const revision = Number(revisionParam)
-    if (!Number.isInteger(revision) || revision < 1) {
-      throw new UnprocessableEntityException(`Invalid revision "${revisionParam}"`)
-    }
-
-    const connectScope = getRequiredConnectScope(request)
-
-    const targetSettings = await this.agentSettingsService.get({
-      connectScope,
-      agentId: request.agent.id,
-      revision,
-    })
-    if (!targetSettings) {
-      throw new NotFoundException(
-        `Revision ${revision} not found for agent with id ${request.agent.id}`,
-      )
-    }
-    const archived = await this.agentSettingsService.archive({
-      connectScope,
-      agentId: request.agent.id,
-      revision,
-    })
-    if (!archived || !archived.success) {
-      throw new UnprocessableEntityException(
-        `Unable to archive revision ${revision} for agent with id ${request.agent.id}`,
-      )
     }
     return { data: { success: true } }
   }
@@ -335,56 +206,5 @@ function toAgentSubAgentDto(entity: AgentSubAgent): AgentSubAgentDto {
       : undefined,
     createdAt: entity.createdAt.getTime(),
     updatedAt: entity.updatedAt.getTime(),
-  }
-}
-
-function toAgentDto({
-  agent,
-  agentSettings,
-}: {
-  agent: Agent
-  agentSettings: AgentSettings
-}): AgentDto {
-  return {
-    createdAt: agent.createdAt.getTime(),
-    greetingMessage: agentSettings.greetingMessage ?? undefined,
-    instructions: agentSettings.instructions,
-    hasCategories: (agent.sessionCategories?.length ?? 0) > 0,
-    id: agent.id,
-    revision: agentSettings.revision,
-    revisionName: agentSettings.revisionName ?? "",
-    revisionDesc: agentSettings.revisionDesc ?? "",
-    isDraft: agentSettings.isDraft,
-    isArchived: agentSettings.isArchived,
-    locale: agentSettings.locale,
-    model: agentSettings.model,
-    name: agent.name,
-    outputJsonSchema: (agentSettings.outputJsonSchema as AgentDto["outputJsonSchema"]) ?? undefined,
-    projectId: agent.projectId,
-    temperature: Number(agentSettings.temperature),
-    type: agent.type,
-    updatedAt: agentSettings.updatedAt.getTime(),
-    documentTagIds: agent.documentTags?.map((tag) => tag.id) || [],
-    resourceLibraryIds: agent.resourceLibraries?.map((library) => library.id) || [],
-    documentsRagMode: agentSettings.documentsRagMode,
-    fillFormEnabled: agentSettings.fillFormEnabled,
-    projectAgentSessionCategoryIds: (agent.sessionCategories ?? [])
-      .map((category) => category.projectAgentSessionCategoryId)
-      .filter(
-        (projectAgentSessionCategoryId): projectAgentSessionCategoryId is string =>
-          projectAgentSessionCategoryId !== null,
-      ),
-    usedProjectAgentSessionCategoryIds: (agent.sessionCategories ?? [])
-      .filter((category) => (category.conversationSessionCategories?.length ?? 0) > 0)
-      .map((category) => category.projectAgentSessionCategoryId)
-      .filter(
-        (projectAgentSessionCategoryId): projectAgentSessionCategoryId is string =>
-          projectAgentSessionCategoryId !== null,
-      ),
-    mcpServers: (agent.agentMcpServers ?? []).map((agentMcpServer) => ({
-      id: agentMcpServer.mcpServer.id,
-      name: agentMcpServer.mcpServer.name,
-      enabled: agentMcpServer.enabled,
-    })),
   }
 }
