+++ apps/api/src/domains/agents/agents.module.ts
@@ -36,6 +36,7 @@ import { ProjectAgentSessionCategoriesGuard } from "./session-categories/project
 import { ProjectAgentSessionCategoriesService } from "./session-categories/project-agent-session-categories.service"
 import { ProjectAgentSessionCategory } from "./session-categories/project-agent-session-category.entity"
 import { AgentSettings } from "./settings/agent-settings.entity"
+import { AgentSettingsController } from "./settings/agent-settings.controller"
 import { AgentSettingsService } from "./settings/agent-settings.service"
 import { AgentSubAgent } from "./sub-agents/agent-sub-agent.entity"
 import { AgentSubAgentsService } from "./sub-agents/agent-sub-agents.service"
@@ -85,6 +86,7 @@ import { AgentSubAgentsService } from "./sub-agents/agent-sub-agents.service"
   ],
   controllers: [
     AgentsController,
+    AgentSettingsController,
     AgentMembershipsController,
     ProjectAgentSessionCategoriesController,
   ],
